ENGR00320248 Wifi-Display UIBC supports mouse & different resolution
Rewrite the low layer UIBC event module. Use the Android framework
event injection to replace the original injecting event to linux device files.
Add an interface "onUibcData" to pass the uibc data from the WifiDisplaySource
into the WifiDisplayController. So that we can inject the event in Java layer.
Added the UIBC scroll infor packet handler in WifiDisplaySource also.
For the different resolution the source calculate the scale and the black block
for the encoded video resolution. When the UIBC coordinate is received it will be
calculated to the local ones via those parameters.

Signed-off-by: Wang Haoran <b50027@freescale.com>